+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/common/model/SwapAsset.kt
@@ -1,11 +1,10 @@
 package co.electriccoin.zcash.ui.common.model
 
-import co.electriccoin.zcash.ui.common.model.near.NearTokenDto
 import co.electriccoin.zcash.ui.design.util.ImageResource
 import co.electriccoin.zcash.ui.design.util.StringResource
 import java.math.BigDecimal
 
-sealed interface SwapAsset {
+interface SwapAsset {
     val tokenTicker: String
     val tokenName: StringResource
     val tokenIcon: ImageResource
@@ -32,14 +31,3 @@ fun SwapAsset.isSame(
 val SwapAsset.isZCashAsset: Boolean
     get() = isSame(token = "zec", chain = "zec")
 
-data class NearSwapAsset(
-    override val tokenName: StringResource,
-    override val tokenIcon: ImageResource,
-    override val blockchain: SwapBlockchain,
-    private val dto: NearTokenDto,
-) : SwapAsset {
-    override val tokenTicker = dto.symbol
-    override val usdPrice = dto.price
-    override val assetId = dto.assetId
-    override val decimals = dto.decimals
-}
