+++ ui-lib/src/test/java/co/electriccoin/zcash/ui/common/model/SwapAssetTestFixture.kt
@@ -1,5 +1,6 @@
 package co.electriccoin.zcash.ui.common.model
 
+import co.electriccoin.zcash.ui.common.model.near.NearSwapAsset
 import co.electriccoin.zcash.ui.common.model.near.NearTokenDto
 import co.electriccoin.zcash.ui.common.repository.SwapAssetsData
 import co.electriccoin.zcash.ui.design.util.imageRes
