+++ ui-lib/src/test/java/co/electriccoin/zcash/ui/common/model/near/NearSwapQuoteValidationTest.kt
@@ -1,6 +1,5 @@
 package co.electriccoin.zcash.ui.common.model.near
 
-import co.electriccoin.zcash.ui.common.model.NearSwapAsset
 import co.electriccoin.zcash.ui.common.model.SwapBlockchain
 import co.electriccoin.zcash.ui.design.util.StringResource
 import co.electriccoin.zcash.ui.design.util.imageRes
